+++ service_crategen/src/commands/generate/codegen/error_types.rs
@@ -108,7 +108,8 @@ pub trait GenerateErrorTypes {
         enum_types.push("/// An error occurred dispatching the HTTP request\nHttpDispatch(HttpDispatchError)".to_string());
         enum_types.push("/// An error was encountered with AWS credentials.\nCredentials(CredentialsError)".to_string());
         enum_types.push("/// A validation error occurred.  Details from AWS are provided.\nValidation(String)".to_string());
-        enum_types.push("/// An unknown error occurred.  The raw HTTP response is provided.\nUnknown(String)".to_string());
+        enum_types.push("/// An error occurred parsing the response payload.\nParseError(String)".to_string());
+        enum_types.push("/// An unknown error occurred.  The raw HTTP response is provided.\nUnknown(BufferedHttpResponse)".to_string());
         Some(enum_types.join(","))
     }
 
@@ -139,7 +140,9 @@ pub trait GenerateErrorTypes {
         type_matchers.push(format!("{error_type}::Credentials(ref err) => err.description()",
                                    error_type = error_type));
         type_matchers.push(format!("{error_type}::HttpDispatch(ref dispatch_error) => dispatch_error.description()", error_type = error_type));
-        type_matchers.push(format!("{error_type}::Unknown(ref cause) => cause",
+        type_matchers.push(format!("{error_type}::ParseError(ref cause) => cause",
+                                   error_type = error_type));
+        type_matchers.push(format!("{error_type}::Unknown(_) => \"unknown error\"",
                                    error_type = error_type));
         Some(type_matchers.join(",\n"))
     }
@@ -155,18 +158,18 @@ impl GenerateErrorTypes for XmlErrorTypes {
     fn generate_error_from_body_impl(&self, operation_name: &str, operation: &Operation, service: &Service) -> String {
         format!("
                 impl {type_name} {{
-                    pub fn from_body(body: &str) -> {type_name} {{
-                        let reader = EventReader::new(body.as_bytes());
-                        let mut stack = XmlResponse::new(reader.into_iter().peekable());
-                        find_start_element(&mut stack);
-                        match Self::deserialize(&mut stack) {{
-                            Ok(parsed_error) => {{
+                    pub fn from_response(res: BufferedHttpResponse) -> {type_name} {{
+                        {{
+                            let reader = EventReader::new(res.body.as_slice());
+                            let mut stack = XmlResponse::new(reader.into_iter().peekable());
+                            find_start_element(&mut stack);
+                            if let Ok(parsed_error) = Self::deserialize(&mut stack) {{
                                 match &parsed_error.code[..] {{
                                     {type_matchers}
                                 }}
-                           }},
-                           Err(_) => {type_name}::Unknown(body.to_string())
+                           }}
                        }}
+                       {type_name}::Unknown(res)
                     }}
 
                     fn deserialize<T>(stack: &mut T) -> Result<XmlError, XmlParseError> where T: Peek + Next {{
@@ -183,7 +186,7 @@ impl GenerateErrorTypes for XmlErrorTypes {
                 impl From<XmlParseError> for {type_name} {{
                     fn from(err: XmlParseError) -> {type_name} {{
                         let XmlParseError(message) = err;
-                        {type_name}::Unknown(message.to_string())
+                        {type_name}::ParseError(message.to_string())
                     }}
                 }}",
                 type_name = error_type_name(service, operation_name))
@@ -220,15 +223,14 @@ impl XmlErrorTypes {
             for error in operation.errors() {
                 let shape = service.get_shape(&error.shape).unwrap();
                 let error_code = shape.error.as_ref().and_then(|http_error| http_error.code.as_ref()).unwrap_or(&error.shape);
-                type_matchers.push(format!("\"{error_code}\" => {error_type}::{error_name}(String::from(parsed_error.message))",
+                type_matchers.push(format!("\"{error_code}\" => return {error_type}::{error_name}(String::from(parsed_error.message))",
                     error_code = error_code,
                     error_type = error_type,
                     error_name = error.idiomatic_error_name()))
             }
         }
 
-        type_matchers.push(format!("_ => {error_type}::Unknown(String::from(body))",
-                                   error_type = error_type));
+        type_matchers.push(format!("_ => {{}}"));
         type_matchers.join(",")
     }
 }
@@ -237,21 +239,19 @@ impl GenerateErrorTypes for JsonErrorTypes {
     fn generate_error_from_body_impl(&self, operation_name: &str, operation: &Operation, service: &Service) -> String {
         format!("
                 impl {type_name} {{
-                    pub fn from_body(body: &str) -> {type_name} {{
-                        match from_str::<SerdeJsonValue>(body) {{
-                            Ok(json) => {{
-                                let raw_error_type = json.get(\"__type\").and_then(|e| e.as_str()).unwrap_or(\"Unknown\");
-                                let error_message = json.get(\"message\").and_then(|m| m.as_str()).unwrap_or(body);
+                    pub fn from_response(res: BufferedHttpResponse) -> {type_name} {{
+                        if let Ok(json) = from_slice::<SerdeJsonValue>(&res.body) {{
+                            let raw_error_type = json.get(\"__type\").and_then(|e| e.as_str()).unwrap_or(\"Unknown\");
+                            let error_message = json.get(\"message\").and_then(|m| m.as_str()).unwrap_or(\"\");
 
-                                let pieces: Vec<&str> = raw_error_type.split(\"#\").collect();
-                                let error_type = pieces.last().expect(\"Expected error type\");
+                            let pieces: Vec<&str> = raw_error_type.split(\"#\").collect();
+                            let error_type = pieces.last().expect(\"Expected error type\");
 
-                                match *error_type {{
-                                    {type_matchers}
-                                }}
-                            }},
-                            Err(_) => {type_name}::Unknown(String::from(body))
+                            match *error_type {{
+                                {type_matchers}
+                            }}
                         }}
+                        return {type_name}::Unknown(res);
                     }}
                 }}",
                 type_name = error_type_name(service, operation_name),
@@ -262,7 +262,7 @@ impl GenerateErrorTypes for JsonErrorTypes {
         format!("
                 impl From<serde_json::error::Error> for {type_name} {{
                     fn from(err: serde_json::error::Error) -> {type_name} {{
-                        {type_name}::Unknown(err.description().to_string())
+                        {type_name}::ParseError(err.description().to_string())
                     }}
                 }}",
                 type_name = error_type_name(service, operation_name))
@@ -279,16 +279,15 @@ impl JsonErrorTypes {
         if operation.errors.is_some() {
             for error in operation.errors() {
                 if error.shape != "ValidationException" {
-                    type_matchers.push(format!("\"{error_shape}\" => {error_type}::{error_name}(String::from(error_message))",
+                    type_matchers.push(format!("\"{error_shape}\" => return {error_type}::{error_name}(String::from(error_message))",
                         error_shape = error.shape,
                         error_type = error_type,
                         error_name = error.idiomatic_error_name()))
                 }
             }
         }
-        type_matchers.push(format!("\"ValidationException\" => {error_type}::Validation(error_message.to_string())", error_type = error_type));
-        type_matchers.push(format!("_ => {error_type}::Unknown(String::from(body))",
-                                   error_type = error_type));
+        type_matchers.push(format!("\"ValidationException\" => return {error_type}::Validation(error_message.to_string())", error_type = error_type));
+        type_matchers.push(format!("_ => {{}}"));
         type_matchers.join(",\n")
     }
 }